pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "daily-report-bot"
version = "0.1.0"
description = "PC作業自動記録・日報生成システム"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "okira" }]
dependencies = [
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Phase 1: MVP
phase1 = [
    "notion-client>=2.2.0",
]
# Phase 3: LLM Integration
llm = [
    "google-genai>=0.3.0",
]
# Windows specific
windows = [
    "pywin32>=306",
    "Pillow>=10.0.0",
    "winocr>=0.0.2",
    "win10toast>=0.9",
]
# Development
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]
# All dependencies
all = [
    "daily-report-bot[phase1,llm,windows,dev]",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
